--- server/utils/helpers/getUserAgent.js
@@ -1,10 +1,0 @@
-const { getDeploymentVersion } = require("../../endpoints/utils");
-
-function getAnythingLLMUserAgent() {
-  const version = getDeploymentVersion() || "VERSION_NOT_FOUND";
-  return `AnythingLLM/${version}`;
-}
-
-module.exports = {
-  getAnythingLLMUserAgent,
-};
